server/requirements.txt
openenv-core[core]==0.2.1


